re #392 - fixed re #393 - fixed re #394 - fixed re #395 - fixed re #396 - fixed re #397 - fixed re #398 - fixed re #399 - fixed re #400 - fixed re #401 - fixed re #402 - fixed re #403 - fixed re #404 - fixed

+++ ios/My Job Pitch-Swift/MyJobPitch/Controllers/BusinessListController.swift
@@ -11,7 +11,12 @@ import MGSwipeTableCell
 
 class BusinessListController: MJPController {
     
+    static var reloadRequest = false
+    
     @IBOutlet weak var tableView: UITableView!
+    @IBOutlet weak var emptyView: UIView!
+    @IBOutlet weak var emptyMessage: UILabel!
+    @IBOutlet weak var emptyButton: UIButton!
     
     var data: NSMutableArray!
     
@@ -23,6 +28,18 @@ class BusinessListController: MJPController {
         data = NSMutableArray()
         refresh()
         
+        if !AppData.user.canCreateBusinesses {
+            tableView.isScrollEnabled = false
+        }
+        
+    }
+    
+    override func viewWillAppear(_ animated: Bool) {
+        super.viewWillAppear(animated)
+        if BusinessListController.reloadRequest {
+            BusinessListController.reloadRequest = false
+            tableView.reloadData()
+        }
     }
     
     func refresh() {
@@ -32,24 +49,38 @@ class BusinessListController: MJPController {
         API.shared().loadBusinesses(success: { (data) in
             AppHelper.hideLoading()
             self.data = data.mutableCopy() as! NSMutableArray
-            
-            if !AppData.user.canCreateBusinesses && self.data.count > 0 {
-                SideMenuController.pushController(id: "locations")
-            } else {
-                self.tableView.reloadData()
-            }
-            
+            self.tableView.reloadData()
+            self.updateUI()
+            AppData.user.businesses = self.data
         }) { (message, errors) in
             self.handleErrors(message: message, errors: errors)
         }
         
     }
     
+    func updateUI() {
+        
+        self.emptyView.isHidden = AppData.user.canCreateBusinesses && self.data.count > 0
+        if self.data.count == 0 {
+            self.emptyMessage.text = "You have not added any\n businesses yet."
+            self.emptyButton.setTitle("Create business", for: .normal)
+        } else if !AppData.user.canCreateBusinesses {
+            self.emptyMessage.text = "Have more than one company?\n Get in touch!"
+            self.emptyButton.setTitle("sales@myjobpitch.com", for: .normal)
+        }
+        
+    }
+    
     @IBAction func addAction(_ sender: Any) {
         
-        let controller = AppHelper.mainStoryboard.instantiateViewController(withIdentifier: "BusinessEdit") as! BusinessEditController
-        controller.savedBusiness = refresh
-        navigationController?.pushViewController(controller, animated: true)
+        if AppData.user.canCreateBusinesses || data.count == 0 {
+            let controller = AppHelper.mainStoryboard.instantiateViewController(withIdentifier: "BusinessEdit") as! BusinessEditController
+            controller.savedBusiness = refresh
+            navigationController?.pushViewController(controller, animated: true)
+        } else {
+            let url = URL(string: "mailto:sales@myjobpitch.com")!
+            UIApplication.shared.openURL(url)
+        }
         
     }
     
@@ -64,7 +95,6 @@ extension BusinessListController: UITableViewDataSource {
     func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
         
         let business = data[indexPath.row] as! Business
-        
         let cell = tableView.dequeueReusableCell(withIdentifier: "BusinessCell", for: indexPath) as! BusinessCell
         
         cell.setData(business)
@@ -85,36 +115,41 @@ extension BusinessListController: UITableViewDataSource {
             })
         ]
         
-        cell.rightButtons = [
-            MGSwipeButton(title: "",
-                          icon: UIImage(named: "delete-big-icon"),
-                          backgroundColor: AppData.yellowColor,
-                          padding: 20,
-                          callback: { (cell) -> Bool in
-                            
-                            let message = String(format: "Are you sure you want to delete %@", business.name)
-                            PopupController.showYellow(message, ok: "Delete", okCallback: {
+        if AppData.user.canCreateBusinesses {
+            
+            cell.rightButtons = [
+                MGSwipeButton(title: "",
+                              icon: UIImage(named: "delete-big-icon"),
+                              backgroundColor: AppData.yellowColor,
+                              padding: 20,
+                              callback: { (cell) -> Bool in
                                 
-                                AppHelper.showLoading("Deleting...")
+                                let message = String(format: "Are you sure you want to delete %@", business.name)
+                                PopupController.showYellow(message, ok: "Delete", okCallback: {
+                                    
+                                    AppHelper.showLoading("Deleting...")
+                                    
+                                    API.shared().deleteBusiness(id: business.id, success: {
+                                        AppHelper.hideLoading()
+                                        self.data.remove(business)
+                                        self.tableView.reloadData()
+                                        self.updateUI()
+                                    }) { (message, errors) in
+                                        self.handleErrors(message: message, errors: errors)
+                                    }
+                                    
+                                    cell.hideSwipe(animated: true)
+                                    
+                                }, cancel: "Cancel", cancelCallback: {
+                                    cell.hideSwipe(animated: true)
+                                })
                                 
-                                API.shared().deleteBusiness(id: business.id, success: {
-                                    AppHelper.hideLoading()
-                                    self.data.remove(business)
-                                    self.tableView.reloadData()
-                                }) { (message, errors) in
-                                    self.handleErrors(message: message, errors: errors)
-                                }
-                                
-                                cell.hideSwipe(animated: true)
-                                
-                            }, cancel: "Cancel", cancelCallback: {
-                                cell.hideSwipe(animated: true)
-                            })
-                            
-                            return false
-            })
-        ]
-
+                                return false
+                })
+            ]
+            
+        }
+        
         return cell
         
     }
@@ -125,11 +160,13 @@ extension BusinessListController: UITableViewDelegate {
     
     func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
         
-        let business = data[indexPath.row] as! Business
-        
-        let controller = AppHelper.mainStoryboard.instantiateViewController(withIdentifier: "LocationList") as! LocationListController
-        controller.business = business
-        navigationController?.pushViewController(controller, animated: true)
+        if indexPath.row < data.count {
+            
+            let controller = AppHelper.mainStoryboard.instantiateViewController(withIdentifier: "LocationList") as! LocationListController
+            controller.business = data[indexPath.row] as! Business
+            navigationController?.pushViewController(controller, animated: true)
+            
+        }
         
     }
